6_langchain-ai/2_langgraph/6_agents/2b_supervision/pyproject.toml
[project]
name = "2-supervision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "langchain[openai]>=1.0.2",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=1.0.2",
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "wolframalpha>=5.1.3",
    "yfinance>=0.2.61",
    "devtools>=0.12.2"
]
